Move test to v1 package
